Filter Panel: Eyedropper Update & Moved Outside of the ColorPicker
See: #9465
See also: #9488
