+++ packages/story-editor/src/components/eyedropper/index.js
